requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2
numpy==1.26.4
pillow==10.4.0
opencv-python-headless==4.10.0.84
scikit-learn==1.5.2
scikit-image==0.24.0
transformers==4.45.2
torch>=2.2.0
torchvision>=0.17.0
timm==1.0.9
matplotlib==3.9.2
requests==2.32.3
diffusers==0.30.3
accelerate==0.34.2
safetensors==0.4.5
tqdm==4.66.5